# docs/reference/openapi_v1.yaml
openapi: 3.0.3
info:
  title: XZe API
  description: AI-powered documentation generator API - Version 1
  contact:
    name: XZe Project
    url: https://github.com/xbcsmith/xze
  license:
    name: MIT OR Apache-2.0
    url: https://github.com/xbcsmith/xze
  version: 1.0.0
servers:
- url: /api/v1
  description: API v1 base path
paths:
...
  /api/v1/search:
    get:
      tags:
      - search
      summary: Handler for semantic search
      operationId: handle_search
      parameters:
      - name: q
        in: query
        description: Search query text
        required: true
        schema:
          type: string
      - name: max_results
        in: query
        description: 'Maximum number of results (default: 10)'
        required: false
        schema:
          type: integer
          nullable: true
          minimum: 0
      - name: min_similarity
        in: query
        description: 'Minimum similarity threshold 0.0-1.0 (default: 0.0)'
        required: false
        schema:
          type: number
          format: float
          nullable: true
      - name: category
        in: query
        description: Filter by documentation category (tutorial, how-to, reference, explanation)
        required: false
        schema:
          type: string
          nullable: true
      responses:
        '200':
          description: Search results with similarity scores
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchErrorResponse'
        '500':
          description: Internal search error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchErrorResponse'
        '502':
          description: Failed to generate embedding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchErrorResponse'
...
components:
  schemas:
...
    SearchConfigResponse:
      type: object
      description: Search configuration details
      required:
      - max_results
      - min_similarity
      properties:
        category_filter:
          type: string
          description: Category filter applied
          example: tutorial
          nullable: true
        max_results:
          type: integer
          description: Maximum results requested
          example: 10
          minimum: 0
        min_similarity:
          type: number
          format: float
          description: Minimum similarity threshold applied
          example: 0.0
    SearchErrorResponse:
      type: object
      description: Search error response
      required:
      - error
      properties:
        details:
          type: string
          description: Additional error details
          nullable: true
        error:
          type: string
          description: Error message
          example: Query string cannot be empty
...
    SearchResponse:
      type: object
      description: Search response with results
      required:
      - query
      - results
      - total_results
      - config
      properties:
        config:
          $ref: '#/components/schemas/SearchConfigResponse'
        query:
          type: string
          description: Original search query
          example: how to implement authentication
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultItem'
          description: List of search results ordered by similarity
        total_results:
          type: integer
          description: Total number of results found
          example: 5
          minimum: 0
    SearchResultItem:
      type: object
      description: Individual search result item
      required:
      - id
      - source_file
      - content
      - similarity
      - chunk_index
      - total_chunks
      - sentence_range
      - avg_chunk_similarity
      properties:
        avg_chunk_similarity:
          type: number
          format: float
          description: Average similarity across all chunks in document
          example: 0.82
        category:
          type: string
          description: Documentation category
          example: how-to
          nullable: true
        chunk_index:
          type: integer
          format: int32
          description: Index of this chunk within the document
          example: 2
        content:
          type: string
          description: Content text of the matching chunk
          example: To implement authentication, first configure the JWT middleware...
        id:
          type: integer
          format: int64
          description: Unique identifier for this document chunk
        sentence_range:
          type: array
          items:
            allOf:
            - type: integer
              format: int32
            - type: integer
              format: int32
          description: Range of sentences in this chunk (start, end)
        similarity:
          type: number
          format: float
          description: Similarity score (0.0 to 1.0, higher is better)
          example: 0.87
        source_file:
          type: string
          description: Source file path
          example: docs/how_to/authentication.md
        title:
          type: string
          description: Document title
          example: Authentication Guide
          nullable: true
        total_chunks:
          type: integer
          format: int32
          description: Total number of chunks in the document
          example: 5
...
  securitySchemes:
    api_key:
      type: apiKey
      in: header
      name: X-API-Key
...